src/scripts/requirements.txt
python-dotenv==1.0.0
pydantic>=2.0.0
requests==2.31.0
rich==13.7.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
aiohttp==3.9.1
flipside==2.0.7
schedule==1.2.1 

# Database
supabase>=2.3.0,<3.0.0
httpx>=0.24.0,<0.25.0  # Pin httpx version to be compatible with supabase
psycopg2-binary==2.9.9
redis==5.0.1

# Data processing
pandas>=2.2.0
numpy==1.26.2
tabulate>=0.9.0
matplotlib>=3.8.0

# API Clients
python-dateutil==2.8.2
hyperliquid-python-sdk>=0.9.0

# Task Scheduling
schedule==1.2.1

# Monitoring & Logging
rich>=13.7.0
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
black==23.12.1
mypy==1.7.1

# Additional dependencies
ccxt>=4.1.13
gotrue>=1.3.0
postgrest>=0.13.0

# Streamlit
streamlit>=1.31.0

# Plotly
plotly>=5.18.0

# YAML
pyyaml>=6.0.1 